audit 2026-04-19: L41 clean-warning sweep + include-cleaner pass
Closes the last deferred item from the 2026-04-19 audit — the hardened
-Wformat=2 / -Wshadow / -Wnull-dereference / -Wconversion /
-Wsign-conversion flag set. Engine and full build now compile with
zero warnings, zero errors; 1883 tests still pass.

L41 root-cause fixes (high-leverage):
- engine/CMakeLists.txt — promote glm-header-only to SYSTEM via CMake
  3.25 SYSTEM property; the prior target_include_directories(SYSTEM)
  re-export was ineffective because the INTERFACE target's own -I came
  first. Kills 346 cascaded GLM qualifier.hpp warnings.
- engine/scripting/node_type_registry.h — `= {}` default-member-init
  on inputIndexByName/outputIndexByName kills 136 -Wmissing-field-
  initializers warnings in the six visual-scripting node registration
  tables without touching each call site.
- engine/physics/ragdoll.cpp, editor/panels/texture_viewer_panel.cpp —
  two genuine -Wmaybe-uninitialized fixes (default-init out-params
  before boundary decompose/extract helpers).
- engine/audio/audio_engine.h — swap std::vector<bool> for
  std::vector<uint8_t>; sidesteps the GCC 15 libstdc++ stl_bvector
  -Warray-bounds false-positive in resize() and avoids proxy-ref
  oddities.
- 26 source files, 148 in-project -Wsign-conversion fixes: size_t
  loop aliases in hot paths, static_cast<size_t>/GLuint/GLsizeiptr
  /GLenum at indexing/API call sites.

Include-cleaner pass:
- 82 unused #include directives removed from 78 engine/*.cpp files.
- 14 flagged removals reverted as false positives (glm/gtc/matrix_
  transform.hpp for unqualified glm::lookAt/perspective/translate;
  ft2build.h for the FT_FREETYPE_H macro in font.cpp).

clang-only fix:
- editor/panels/model_viewer_panel.cpp:670-671 — explicit static_cast
  <float> on (int) / 10.0f truncate-to-one-decimal to silence
  clang -Wimplicit-int-float-conversion while preserving the
  intentional truncation.

+++ engine/animation/skeleton.cpp
@@ -17,7 +17,7 @@ int Skeleton::findJoint(const std::string& name) const
 {
     for (int i = 0; i < static_cast<int>(m_joints.size()); ++i)
     {
-        if (m_joints[i].name == name)
+        if (m_joints[static_cast<size_t>(i)].name == name)
         {
             return i;
         }
